requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.15.0
Django==3.2
django-cors-headers==3.10.1
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
dnspython==2.1.0
gunicorn==20.1.0
install==1.3.5
psycopg2==2.9.3
pycparser==2.21
pytz==2021.3
six==1.16.0
sqlparse==0.2.4
whitenoise==5.3.0
